Change the way injection works.
Basically, ApplicationHandler (Jersey's) creates a ServiceLocator,
(HK2 counterpart of Guice's Injector) and starts injecting
components straight away upon construction.

But this least to a problem, as we can't tie Guice to it until
we have a ServiceLocator (after construction and initialization of
singletons).

Luckily we can specify an extra HK2 binder upon construction of
the ApplicationHandler, therefore the GuiceBinder simply binds
every single key it finds in a Guice injector to a simple HK2
factory (counterpart of Provider) fetching instances from
Guice's injector.

This seems to do the trick, not happy about it, but it works.